API/requirements.txt
flask
flask_cors
pyOpenSSL
